fix: Move MinecraftEncoderProxy entirely to the CommonChannelInjector to elegantly handle the Netty type parameter matching.
